renaming XmlLexer→LexingXmlStreamReader, XmlScanner→XmlLexer, ScannerState→LexerState
